requirements.txt
click==8.1.3
PyYAML==6.0
termcolor==2.3.0
